% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calendar}
\alias{calendar}
\title{Calendar}
\format{
Can be re-obtained from the image
}
\usage{
data("calendar")
}
\description{
A \code{\link{data.frame}} for the calendar
}
\keyword{datasets}
