% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneNames}
\alias{geneNames}
\title{A character vector containing gene names of filtered genes.}
\format{An object of class \code{character} of length 10549.}
\usage{
geneNames
}
\description{
A character vector containing gene names of filtered genes.
}
\keyword{datasets}
