% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_raw_data.R
\name{importRawData}
\alias{importRawData}
\title{Download and import the raw dataset.}
\usage{
importRawData()
}
\description{
Download the dataset from GEO, filter, and create a
\code{SingleCellExperiment object}
}
