% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_score.R
\name{proximity_score}
\alias{proximity_score}
\title{Proximity score.}
\usage{
proximity_score(rd, cl, k = 10, smooth = k)
}
\arguments{
\item{rd}{The reduced dimension matrix of the cells}

\item{cl}{the vector of conditions}

\item{k}{The number of neighbours to consider when computing the score.
Default to 10.}

\item{smooth}{The smoothing parameter. Default to k. Lower values mean that
we smooth more.}
}
\description{
Compute a proximity score to show whether nearby cells have the
same condition of not
}
\examples{
sd <- create_differential_topology(n_cells = 200, shift = 0,
                                   unbalance_level = .8)
scores <- proximity_score(sd$rd, sd$cl, 4)
ggplot(data.frame(sd$rd, scores = scores$scaled_scores),
       aes(x = Dim1, y = Dim2, col = scores)) +
  geom_point() +
  theme_classic()
}
