% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{The object}
\format{
Can be re-obtained by runing first chunk of the vignette.
}
\usage{
data(sce)
}
\description{
A \code{\link{SingleCellExperiment}} object of 9268 cells, derived from [McFaline-Figueroa, et al.](https://www.nature.com/articles/s41588-019-0489-5)
}
\keyword{datasets}
